modified:   .gitignore 	modified:   new_pipeline/runtime/combined_loss.py

+++ new_pipeline/runtime/combined_loss.py
@@ -36,8 +36,8 @@ class CombinedLoss(nn.Module):
         for fn, weight in self.weighted_losses:
             # Check if it's a standard NN loss (like CE) or an SMP loss
             if isinstance(fn, (nn.CrossEntropyLoss, nn.NLLLoss)):
-                target_sq = target_sq.float()  # Ensure targets are long for CE loss
-                pred_sq = pred_sq.long()  # Ensure predictions are float for loss calculations
+                target_sq = target_sq.long()  # Ensure targets are long for CE losss
+                pred_sq = pred_sq.long()  # Ensure predictions are long for loss calculation
                 loss += fn(pred_sq, target_sq) * weight
             else:
                 loss += fn(pred_sq, target) * weight
